update utils plugin
