+++ sigma/plugins/utils/echo.py
@@ -1,26 +1,9 @@
-from sigma.plugin import Plugin
-from sigma.utils import create_logger
-
 from config import permitted_id
 
 
-class Echo(Plugin):
-    is_global = True
-    log = create_logger('echo')
-
-    async def on_message(self, message, pfx):
-        if message.content.startswith(pfx + 'echo' + ' '):
-            cmd_name = 'Echo'
-            try:
-                self.log.info('User %s [%s] on server %s [%s], used the ' + cmd_name + ' command on #%s channel',
-                              message.author,
-                              message.author.id, message.server.name, message.server.id, message.channel)
-            except:
-                self.log.info('User %s [%s], used the ' + cmd_name + ' command.',
-                              message.author,
-                              message.author.id)
-            if message.author.id in permitted_id:
-                await self.client.send_message(message.channel, message.content[len('echo') + len(pfx):])
-            else:
-                await self.client.send_message(message.channel,
-                                    'Sorry, <@' + message.author.id + '>, you do not have permission to do that...')
+async def echo(cmd, message, args):
+    if message.author.id in permitted_id:
+        await cmd.reply(' '.join(args))
+    else:
+        msg = 'Sorry, <@{:s}>, you do not have permission to do that...'
+        await cmd.reply(msg.format(message.author.id))
